// ui/src/app/models/SportsClub.ts
export class SportsClub {
  clubName: string | undefined;
  matchesPlayed: string | undefined;
  wins: string | undefined;
  losses: string | undefined;
  draws: string | undefined;
  goalsScored: string | undefined;
  goalsReceived: string | undefined;
  goalDifference: string | undefined;
  points: string | undefined;
}